requirements.txt
# Core dependencies
opencv-python-headless==4.8.1.78  # Lighter version for Raspberry Pi
mediapipe==0.10.8
numpy==1.26.2
Pillow==10.1.0
pygame==2.5.2

# Hardware acceleration
picamera2  # For Raspberry Pi camera module
libcamera  # Camera stack for Raspberry Pi

# System utilities
psutil  # For system monitoring
rpi.gpio  # For Raspberry Pi GPIO control

# Optional: TensorFlow Lite (uncomment if needed)
# tflite-runtime  # Lightweight TensorFlow for edge devices 